+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 - Default forum settings to the admin backend
+- `related_name` to foreign keys that where missing them
 
 ### Changed