+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [In Development] - Unreleased
 
+
+## [1.11.0] - 2022-08-24
+
 ### Added
 
 - Default forum settings to the admin backend